requirements.txt
# Python 依赖
# 核心数据获取
requests>=2.28.0
baostock>=0.8.8

# 数据处理
pandas>=1.5.0
numpy>=1.23.0

# 技术分析（可选，部分指标内置计算）
# ta-lib 需要系统级安装，如无法安装可跳过

# 机器学习模型（可选 — Qlib ML 打分，当前禁用）
# 启用需要独立 venv + qlib 数据，详见 .github/copilot-instructions.md
# qlib>=0.9.0
# lightgbm>=3.3.0
# scikit-learn>=1.1.0

# FinRL（可选，PPO模型）
# stable-baselines3>=1.7.0

# 调度
schedule>=1.1.0

# 环境配置
python-dotenv>=0.20.0

# 日志增强（可选）
# loguru>=0.6.0
